+++ src/planning/bezier_smooth.h
@@ -3,5 +3,5 @@
 namespace planning {
 	void GetCurveControlPoints(point_arr knots, point_arr &firstControlPoints, point_arr &secondControlPoints);
 	std::vector<double> GetFirstControlPoints(std::vector<double> rhs);
-	geometry_msgs::Point bezier_point(int u, point_arr ctrl_points);
+	geometry_msgs::Point bezier_point(float u, point_arr ctrl_points);
 }
